Extract common provider implementation to a base class
Most providers (except FileSystem) were near exact implementations.
FileSystem is still similar, but retains its own implementation to
handle non-HTTP fetched files.
